# src/app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API MemoBelc",
    "description": "Documentação da API MemoBelc. Use **Authorize** para informar o token Bearer e testar rotas protegidas."
  },
  "host": "127.0.0.1:5000",
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Token JWT: use o valor 'Bearer <seu_token>'"
    }
  },
  "security": [],
...
  "paths": {
...
    "/plans/admin": {
      "get": {
        "tags": ["Plans"],
        "summary": "Listar todos os planos (admin)",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Planos" }, "403": { "description": "Não autorizado" } }
      },
      "post": {
        "tags": ["Plans"],
        "summary": "Criar plano (admin)",
        "security": [{ "Bearer": [] }],
        "parameters": [{
          "name": "body",
          "in": "body",
          "schema": {
            "type": "object",
            "required": ["name", "price"],
            "properties": {
              "name": { "type": "string" },
              "description": { "type": "string" },
              "price": { "type": "number" },
              "cycle": { "type": "string", "enum": ["WEEKLY", "MONTHLY", "QUARTERLY", "SEMIANNUALLY", "YEARLY"] },
              "trial_days": { "type": "integer" },
              "included_service_keys": { "type": "array", "items": { "type": "string" } },
              "included_book_ids": { "type": "array", "items": { "type": "string" } },
              "google_play_product_id": { "type": "string" }
            }
          }
        }],
        "responses": { "201": { "description": "Criado" } }
      }
    },
...
    "/coupons/admin": {
      "get": {
        "tags": ["Coupons"],
        "summary": "Listar cupons (admin)",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Cupons" } }
      },
      "post": {
        "tags": ["Coupons"],
        "summary": "Criar cupom (admin)",
        "security": [{ "Bearer": [] }],
        "responses": { "201": { "description": "Criado" } }
      }
    },
...
    "/bundles/admin": {
      "get": {
        "tags": ["Bundles"],
        "summary": "Listar conjuntos (admin)",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Conjuntos" } }
      },
      "post": {
        "tags": ["Bundles"],
        "summary": "Criar conjunto (admin)",
        "security": [{ "Bearer": [] }],
        "responses": { "201": { "description": "Criado" } }
      }
    },
...
  }
}